// src/lib/types/products.d.ts
export type homeProduct = {
  product_id: string;
  title: string;
  handle: string;
  subcategory_title?: string;
  featured_image_url: string;
  isfeature: boolean;
  isnew: boolean;
  subcategory_path?: string;
  minprice: number;
  category_path?: string;
};
export type ProductModel = {
  id?: string; // UUID generated by the database
  handle: string;
  title: string;
  description: string;
  available_for_sale: boolean;
  featured_image_url: string;
  seo_title: string;
  seo_description: string;
  tags: string[];
  updated_at?: string; // Timestamp generated by the database
  minprice: number;
  maxprice: number;
  
  category: string;
  isNew: boolean;
  isFeatured: boolean;
};
export interface SingelProductModel {
  id: string;
  handle: string;
  title: string;
  description: string;
  available_for_sale: boolean;
  seo_title: string;
  
  seo_description: string;
  tags: string[];
  minprice: number;
  featured_image_url: string;
  isnew: boolean;
  isfeature: boolean;
  maxprice: number;
  product_images: Image[];
  product_options: ProductOption[];
  product_variants: ProductVariant[];
  products_features: ProductsFeatures[];
  product_specifications: ProductSpecification[];
  reviews: ProductReviews[];
}
export type ProductOption = {
  id: string;
  name: string;
  values: string[];
};
export type ProductsFeatures = {
  id: string;
  name: string;
  value: string[];
};
export type ProductSpecification = {
  id: string;
  name: string;
  value: string;
};

export type Image = {
  id: string;
  url: string[];
};

export type SEO = {
  title: string;
  description: string;
};
type ProductVariant = {
  id: string;
  title: string;
  price: number; // ✅ مطلوب
  available_for_sale: boolean;
  variant_selected_options: {
    name: string;
    value: string;
    id: string;
  }[];
};
export type ProductReviews = {
  id: string;
  rating: number;
  comment: string;
  created_at: string;
  profiles: {
    id: string;
    name: string;
    email: string;
  };
};
export type ProductAd = {
  id: string; // uuid
  title: string;
  text2: string;
  text3: string;
  description: string;
  href: string;
  img: string;
  start: number;
  feature: string;
};